Improve naming of protocol, structs and function
