--- Dynavity/Dynavity/view-model/HtmlExporter.swift
@@ -1,5 +1,0 @@
-import Combine
-
-protocol HtmlExporter {
-    func getHtmlPublisher(markupText: String, markupType: MarkupElement.MarkupType) -> AnyPublisher<String, Never>
-}
